Staging/chipnet

+++ fabfile.py
@@ -2,6 +2,15 @@ from patchwork.transfers import rsync
 from fabric import task
 from fabric.connection import Connection
 from dotenv import dotenv_values
+from pathlib import Path
+
+
+def get_active_branch_name():
+    head_dir = Path(".") / ".git" / "HEAD"
+    with head_dir.open("r") as f: content = f.read().splitlines()
+    for line in content:
+        if line[0:4] == "ref:":
+            return line.partition("refs/heads/")[2]
 
 
 config = dotenv_values(".env")
@@ -21,22 +30,27 @@ def chipnet(ctx):
 
 @task
 def mainnet(ctx):
-    ctx.config.network = 'mainnet'
-    ctx.config.project_dir = f'/root/{project}'
-    ctx.config.run.env['conn'] = Connection(
-        config['MAINNET_SERVER_HOST'],
-        user=config['MAINNET_SERVER_USER']
-    )
-
+    branch = get_active_branch_name()
+    if branch == 'master':
+        ctx.config.network = 'mainnet'
+        ctx.config.project_dir = f'/root/{project}'
+        ctx.config.run.env['conn'] = Connection(
+            config['MAINNET_SERVER_HOST'],
+            user=config['MAINNET_SERVER_USER']
+        )
+    else:
+        print(f'\nAborted: {branch} branch is not allowed to access mainnet server!\n')
 
 @task
 def uname(ctx):
+    if 'network' not in ctx.config.keys(): return
     conn = ctx.config.run.env['conn']
     conn.run('uname -a')
 
 
 @task
 def sync(ctx):
+    if 'network' not in ctx.config.keys(): return
     conn = ctx.config.run.env['conn']
     rsync(
         conn,
@@ -59,6 +73,7 @@ def sync(ctx):
 
 @task
 def build(ctx):
+    if 'network' not in ctx.config.keys(): return
     conn = ctx.config.run.env['conn']
     with conn.cd(ctx.config.project_dir):
         conn.run(f'docker-compose -f compose/{ctx.config.network}.yml --env-file {ctx.config.project_dir}/.env build')
@@ -66,6 +81,7 @@ def build(ctx):
 
 @task
 def up(ctx):
+    if 'network' not in ctx.config.keys(): return
     conn = ctx.config.run.env['conn']
     with conn.cd(ctx.config.project_dir):
         conn.run(f'docker-compose -f compose/{ctx.config.network}.yml --env-file {ctx.config.project_dir}/.env up -d')
@@ -73,6 +89,7 @@ def up(ctx):
 
 @task
 def down(ctx):
+    if 'network' not in ctx.config.keys(): return
     conn = ctx.config.run.env['conn']
     with conn.cd(ctx.config.project_dir):
         conn.run(f'docker-compose -f compose/{ctx.config.network}.yml --env-file {ctx.config.project_dir}/.env down --remove-orphans')
@@ -80,6 +97,7 @@ def down(ctx):
 
 @task
 def deploy(ctx):
+    if 'network' not in ctx.config.keys(): return
     sync(ctx)
     build(ctx)
     down(ctx)
@@ -88,6 +106,7 @@ def deploy(ctx):
 
 @task
 def nginx(ctx):
+    if 'network' not in ctx.config.keys(): return
     sync(ctx)
     conn = ctx.config.run.env['conn']
     with conn.cd(f'{ctx.config.project_dir}/compose'):
@@ -108,6 +127,7 @@ def nginx(ctx):
 
 @task
 def logs(ctx):
+    if 'network' not in ctx.config.keys(): return
     conn = ctx.config.run.env['conn']
     with conn.cd(ctx.config.project_dir):
         conn.run(f'docker-compose -f compose/{ctx.config.network}.yml --env-file {ctx.config.project_dir}/.env logs  -f web')
@@ -115,6 +135,7 @@ def logs(ctx):
 
 @task
 def reports(ctx):
+    if 'network' not in ctx.config.keys(): return
     conn = ctx.config.run.env['conn']
     with conn.cd(ctx.config.project_dir):
         conn.run(f'docker-compose -f compose/{ctx.config.network}.yml --env-file {ctx.config.project_dir}/.env exec -T web python manage.py reports -p paytaca')
